pyproject.toml
[project]
name = "chatbot-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.0",
    "langchain[anthropic,openai]>=0.3.27",
    "langgraph>=0.6.3",
    "python-dotenv>=1.1.1",
]
